# app/lib/openapi/model/carPositionDto.ts
/**
 * rent car api
 * rent car api description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressDto } from './addressDto';


export interface CarPositionDto { 
    carId: string;
    address: AddressDto;
    lastUpdate: string;
}